Fixed some compilation warnings + replaced int and unsigned int with int32_t and uint32_t

+++ Example/src/fps_scene/PLayerBehavior.cpp
@@ -41,7 +41,7 @@ void PlayerBehavior::processKeybordInput(float deltaTime) {
 }
 
 void PlayerBehavior::processMouseInput() {
-    double xpos, ypos;
+    float xpos, ypos;
     Input::getCursorPos(&xpos, &ypos);
 
     if (firstMouse) {
